+++ app/src/main/java/com/example/planet/domain/repository/UserRepository.kt
@@ -5,7 +5,7 @@ import com.example.planet.data.remote.dto.response.user.UserInfo
 import kotlinx.coroutines.flow.Flow
 
 interface UserRepository {
-    suspend fun getUserInfo(userId: Int): Flow<ApiState>
+    suspend fun getUserInfo(userId: Long): Flow<ApiState>
     suspend fun putUserInfo(userInfo: UserInfo): Flow<ApiState>
