+++ app/src/main/java/com/example/planet/domain/usecase/plogging/GetPloggingIdUseCase.kt
@@ -7,7 +7,7 @@ import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
 
 class GetPloggingIdUseCase @Inject constructor(private val ploggingRepository: PloggingRepository) {
-    suspend operator fun invoke(userId: Int): Flow<ApiState> {
+    suspend operator fun invoke(userId: Long): Flow<ApiState> {
         return ploggingRepository.getPloggingId(userId)
     }
 }
