+++ app/src/main/java/com/example/planet/domain/usecase/ranking/university/GetMyUniversityRankUseCase.kt
@@ -7,6 +7,6 @@ import javax.inject.Inject
 
 class GetMyUniversityRankUseCase @Inject constructor(private val rankRepository: RankRepository) {
     suspend operator fun invoke(userId: Long): Flow<ApiState> {
-        return rankRepository.getUniversityMyRanking(userId.toInt())
+        return rankRepository.getUniversityMyRanking(userId)
     }
 }
